pyproject.toml
[tool.poetry]
name = "todo-api"
version = "0.1.0"
description = ""
authors = ["SoSaymon <szymon.chirowski@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
graphene = "^3.3"
starlette-graphene3 = "^0.6.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.12"}
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.23"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
